Added contact form
Professor: I wrote this based on a combination of an example found online and my old HTML5 Textbook. I do not have a background in php, and I can't say if this is a good form or a bad form, but I'm hoping we can discuss it after you look at it.